+++ src/main/java/com/diegodeleon/kinalapp/repository/VentaRepository.java
@@ -4,6 +4,6 @@ import com.diegodeleon.kinalapp.entity.Venta;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
-public interface VentaRepository extends JpaRepository<Venta,Integer> {
+public interface VentaRepository extends JpaRepository<Venta,Long> {
     List<Venta> findByEstado(int estado);
 }
